CLI-607 Add abstraction level over SCA scanner runners
