+++ src/cli/commands/analyze/dependency-risk-helpers/sca-scanner-runner-base.ts
@@ -1,0 +1,132 @@
+/*
+ * SonarQube CLI
+ * Copyright (C) SonarSource Sàrl
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+
+import { rmSync } from 'node:fs';
+
+import { LOG_FILE } from '../../../../lib/config-constants.ts';
+import logger from '../../../../lib/logger.ts';
+import type { SpawnResult } from '../../../../lib/process.ts';
+import { warn } from '../../../../ui';
+import { CommandFailedError } from '../../_common/error.ts';
+import { type ScaScannerInstaller } from '../../_common/install/sca-scanner.ts';
+import { type ScaScannerSpawner } from './sca-scanner-spawner.ts';
+
+const REDACTED_TOKEN = '***';
+
+export const SCA_SCANNER_START_FAILURE_HINT =
+  'Verify that the SCA scanner is installed and can run on this machine, then retry.';
+export const SCA_SCANNER_PARSE_FAILURE_HINT = `Inspect ${LOG_FILE} for the raw sca-scanner output, then retry.`;
+export const SCA_SCANNER_EXIT_FAILURE_HINT = `Inspect ${LOG_FILE} for the underlying sca-scanner error, fix the reported issue, then retry.`;
+
+export interface ScaScannerInvocation {
+  baseDir: string;
+  apiBaseUrl: string;
+  downloadBaseUrl: string;
+  sonarToken: string;
+  projectKey: string;
+  cacheDir: string;
+  workDir: string;
+  scannerProperties: Record<string, string>;
+  excludedPaths: string[];
+  includeGitIgnoredPaths: boolean;
+  debug: boolean;
+}
+
+/**
+ * Shared skeleton for runners that drive a `sca-scanner` subcommand: install
+ * the binary, spawn the subcommand, validate the exit code, parse stdout, and
+ * always clean up the work dir.
+ */
+export abstract class ScaScannerRunnerBase<T> {
+  constructor(
+    protected readonly installer: ScaScannerInstaller,
+    protected readonly spawner: ScaScannerSpawner,
+  ) {}
+
+  async run(invocation: ScaScannerInvocation): Promise<T> {
+    const args = this.buildArgs(invocation);
+    logger.debug(`sca-scanner args: ${JSON.stringify(this.redactedArgs(args))}`);
+    const binaryPath = await this.installer.install();
+    try {
+      const result = await this.spawnScanner(binaryPath, args);
+      logger.info(`SCA Scanner stdout\n${result.stdout}`);
+      logger.warn(`SCA Scanner stderr\n${result.stderr}`);
+      this.assertSuccessfulExit(result);
+      return this.parseResult(result);
+    } finally {
+      this.cleanupWorkDir(invocation.workDir);
+    }
+  }
+
+  /** argv for the subcommand this runner drives. */
+  protected abstract buildArgs(invocation: ScaScannerInvocation): string[];
+
+  /** Prefix prepended to every error message (e.g. `Manifest discovery error: failed to parse output (…)`). */
+  protected abstract readonly errorPrefix: string;
+
+  /** Interprets stdout of a successful (exit-code 0) run. */
+  protected abstract parseResult(result: SpawnResult): T;
+
+  private redactedArgs(args: string[]): string[] {
+    return args.map((arg) =>
+      arg.startsWith('--sonar-token=') ? `--sonar-token=${REDACTED_TOKEN}` : arg,
+    );
+  }
+
+  protected parseJson(result: SpawnResult): unknown {
+    try {
+      return JSON.parse(result.stdout);
+    } catch (err) {
+      throw new CommandFailedError(
+        `${this.errorPrefix}: failed to parse output (${(err as Error).message})`,
+        { remediationHint: SCA_SCANNER_PARSE_FAILURE_HINT },
+      );
+    }
+  }
+
+  private async spawnScanner(binaryPath: string, args: string[]): Promise<SpawnResult> {
+    try {
+      return await this.spawner.spawn(binaryPath, args);
+    } catch (err) {
+      throw new CommandFailedError(`${this.errorPrefix}: ${(err as Error).message}`, {
+        remediationHint: SCA_SCANNER_START_FAILURE_HINT,
+      });
+    }
+  }
+
+  private assertSuccessfulExit(result: SpawnResult): void {
+    const exitCode = result.exitCode ?? 1;
+    if (exitCode !== 0) {
+      throw new CommandFailedError(
+        `${this.errorPrefix}: sca-scanner exited with code ${exitCode}.`,
+        { remediationHint: SCA_SCANNER_EXIT_FAILURE_HINT },
+      );
+    }
+  }
+
+  private cleanupWorkDir(workDir: string): void {
+    try {
+      rmSync(workDir, { recursive: true, force: true });
+    } catch (err) {
+      const reason = err instanceof Error ? err.message : String(err);
+      warn(`Failed to clean up SCA scanner working directory ${workDir}: ${reason}`);
+    }
+  }
+}
